# Improve is_palindrome performance significantly (C++03 compat)