--- a/build.jam
+++ b/build.jam
@@ -34,6 +34,7 @@ explicit
     [ alias all : boost_algorithm test
         example
         minmax/example minmax/test
+        performance
         string/example string/test
         ]
     ;