--- a/performance/Jamfile.v2
+++ b/performance/Jamfile.v2
@@ -1,0 +1,30 @@
+#==============================================================================
+#   Copyright (c) Antony Polukhin, 2012-2024.
+#   Copyright (c) James E. King III, 2025.
+#
+#   Distributed under the Boost Software License, Version 1.0. (See accompanying
+#   file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+#==============================================================================
+
+# performance tests
+import testing ;
+import path ;
+
+path-constant TEST_DIR : . ;
+
+project performance
+    : source-location ./
+    : requirements
+      <library>/boost/chrono//boost_chrono
+      <library>/boost/program_options//boost_program_options
+      <library>/boost/system//boost_system
+      <link>static
+      <target-os>freebsd:<linkflags>"-lrt"
+      <target-os>linux:<linkflags>"-lrt"
+      <toolset>gcc:<cxxflags>-fvisibility=hidden
+      <toolset>intel-linux:<cxxflags>-fvisibility=hidden
+      <toolset>sun:<cxxflags>-xldscope=hidden
+    : default-build release
+ ;
+
+run perf_is_palindrome.cpp : $(TEST_DIR) ;